+++ inc/field.class.php
@@ -826,21 +826,19 @@ class PluginFieldsField extends CommonDBChild
         //get fields for this container
         $field_obj = new self();
         $fields    = $field_obj->find(['plugin_fields_containers_id' => $c_id, 'is_active' => 1], 'ranking');
-        echo "<form method='POST' action='" . $CFG_GLPI['root_doc'] . "/plugins/fields/front/container.form.php'>";
+        echo "<form method='POST' class='mt-5' action='" . $CFG_GLPI['root_doc'] . "/plugins/fields/front/container.form.php'>";
         echo Html::hidden('plugin_fields_containers_id', ['value' => $c_id]);
         echo Html::hidden('items_id', ['value' => $item->getID()]);
         echo Html::hidden('itemtype', ['value' => $item->getType()]);
-        echo "<table class='tab_cadre_fixe'>";
         echo self::prepareHtmlFields($fields, $item, $canedit);
 
         if ($canedit) {
-            echo "<tr><td class='tab_bg_2 center' colspan='4'>";
+            echo "<div class='d-flex flex-row-reverse px-4 mb-4'>";
             echo "<input class='btn btn-primary' type='submit' name='update_fields_values' value=\"" .
             _sx('button', 'Save') . "\" class='submit'>";
             echo '</td></tr>';
         }
 
-        echo '</table>';
         Html::closeForm();
 
         return true;
@@ -960,17 +958,15 @@ class PluginFieldsField extends CommonDBChild
             return;
         }
 
+        $class = match (true) {
+            !($item instanceof CommonITILObject) && $item instanceof CommonDropdown => 'card-body row',
+            // @phpstan-ignore-next-line -> Instanceof between CommonDBTM and CommonDropdown will always evaluate to false.
+            !($item instanceof CommonITILObject) && !($item instanceof CommonDropdown) => 'card-body d-flex flex-wrap', // lign 969
+            default => '',
+        };
         $html_id = 'plugin_fields_container_' . mt_rand();
-        if (str_contains($current_url, 'helpdesk.public.php')) {
-            echo "<div id='{$html_id}' class='card-body row mx-0' style='border-top:0'>";
-            echo "<div class='offset-md-1 col-md-8 col-xxl-6'>";
-            $field_options = [
-                'label_class' => 'col-lg-3',
-                'input_class' => 'col-lg-9',
-            ];
-        } else {
-            echo "<div id='{$html_id}'>";
-        }
+
+        echo "<div id='{$html_id}' class='" . $class . "'>";
         $display_condition = new PluginFieldsContainerDisplayCondition();
         if ($display_condition->computeDisplayContainer($item, $c_id)) {
             self::showDomContainer(
@@ -978,11 +974,8 @@ class PluginFieldsField extends CommonDBChild
                 $item,
                 $type,
                 $subtype,
-                $field_options ?? [],
+                [],
             );
-        }
-        if (str_contains($current_url, 'helpdesk.public.php')) {
-            echo '</div>';
         }
         echo '</div>';
 
@@ -1085,6 +1078,7 @@ JAVASCRIPT
         $massiveaction = false,
         $field_options = []
     ) {
+
         if (empty($fields)) {
             return false;
         }