+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## ['UNRELEASED']
+
+### Fixed
+
+- Fix UI with GLPI 11
+- Fix `tab` container not displayed
+
 ## [1.22.0] - 2025-09-30
 
 ### Added